+++ packages/german-adjectives-dict/package.json
@@ -2,7 +2,7 @@
   "name": "german-adjectives-dict",
   "version": "1.1.8",
   "description": "German adjectives dictionnary, based on the german-pos-dict content",
-  "main": "dist/adjectives.json",
+  "main": "dist/index.js",
   "scripts": {
     "clean": "rm -rf dist",
     "test": "nyc --reporter=lcov --reporter=text mocha",
