+++ packages/german-adjectives/example.js
@@ -5,7 +5,7 @@
  */
 
 const GermanAdjectivesLib = require('./dist/index.js');
-const GermanAdjectives = require('german-adjectives-dict');
+const GermanAdjectives = require('german-adjectives-dict/dist/adjectives.json');
 
 // neuen
 console.log(GermanAdjectivesLib.agreeGermanAdjective(null, GermanAdjectives, 'neu', 'DATIVE', 'M', 'S', 'DEFINITE'));
