+++ packages/german-adjectives/lib/index.ts
@@ -4,28 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export interface AdjectiveGenderInfo {
-  P: string;
-  F: string;
-  M: string;
-  N: string;
-}
-
-export interface AdjectiveInfoCase {
-  DEF: AdjectiveGenderInfo;
-  IND: AdjectiveGenderInfo;
-  SOL: AdjectiveGenderInfo;
-}
-
-export interface AdjectiveInfo {
-  AKK: AdjectiveInfoCase;
-  DAT: AdjectiveInfoCase;
-  GEN: AdjectiveInfoCase;
-  NOM: AdjectiveInfoCase;
-}
-export interface AdjectivesInfo {
-  [key: string]: AdjectiveInfo;
-}
+import { AdjectivesInfo, AdjectiveInfo } from 'german-adjectives-dict';
 
 export type GermanCases = 'NOMINATIVE' | 'ACCUSATIVE' | 'DATIVE' | 'GENITIVE';
 export type Genders = 'M' | 'F' | 'N';
